Remove further List uses

+++ include/task.h
@@ -17,7 +17,7 @@ void task_setup();
 void task_teardown();
 task_t *task_create();
 void task_start(task_t *task);
-int task_get_thread_count();
+size_t task_get_thread_count();
 bool task_wait_for_zero_threads();
 void task_run(task_t *task);
 void task_destroy(task_t *task);
